Moved comment reformatting to experimental features
Reformatting user formatted comments seems in general like a bad
direction. Reformatting would basically disable usage of licenses or
ascii-arts in comments which is pretty much deal breaker in most of
the cases. Wrapping long, initially single-line, line_comments sounds
like useful idea and it should be revisited but nevertheless any
changes to comments should be done sparringly.

This feature will stay in purely experimental stage and it will be not
available for the time being until proper limitations are clearly
stated.