Don't treat missing value of one value keyword as an error
Since missing value might lead to StopIteration exception for now
let's just allow that value to be missing. Previously it'd stop
formating of remaining files in the batch. More details: PEP 479

Although it's an issue with incorrect CMake code rather than
parsing/formatting such code nevertheless it'd be more of an annoyance
in that case