Support formatting of custom commands without keyworded arguments
It may happen that custom command doesn't use cmake_parse_arguments
because its definition doesn't intent to use keyworded arguments but
it should still be formatted.

On the other hand it might be so that given custom command doesn't use
cmake_parse_arguments but uses different way to support keyworded or
keyword-like arguments. It's possible to opt-out from reformatting
these with line comment:
# gersemi: ignore
Presence of this line comment will make gersemi ignore that custom
command definition so that it's treated as unknown command. It's still
preferred to just not pass that definition with --definitions
parameter instead.